requirements.txt
fastapi
uvicorn[standard]
python-multipart
langchain
langchain-huggingface
langchain-community
langchain-text-splitters
huggingface-hub
sentence-transformers
requests
faiss-cpu
numpy
pydantic
python-dotenv
pypdf
transformers
torch
scipy
librosa
soundfile
gTTS
pyttsx3